fix: use Angular CLI in shortcuts
